+++ CHANGELOG.rst
@@ -15,9 +15,16 @@ Features
 * Created a route to handle ``/robots.txt`` requests.
 * Created a route to handle ``/humans.txt`` requests.
 * Created Trumbowyg Editor as a standalone Django application.
+* Implemented 500 error handler.
 * Provided command to generate project's build ids.
 * Provided caching configuration fro the project.
 * Provided a way to configure site URL.
+
+
+Improvements
+^^^^^^^^^^^^
+
+* Refactor 404 handler and move it to ``core`` app.
 
 
 Trivial/Internal Changes